+++ example/spec/helpers/testUtils.js
@@ -1,10 +1,16 @@
 const fs = require('fs');
-const { S3 } = require('aws-sdk');
+const {
+  aws: { s3 },
+  stringUtils: { globalReplace }
+} = require('@cumulus/common');
 const { Config } = require('kes');
 const lodash = require('lodash');
 const { exec } = require('child-process-promise');
+const path = require('path');
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000000;
+
+const timestampedTestDataPrefix = (prefix) => `${prefix}-${(new Date().getTime())}-test-data/pdrs`;
 
 /**
  * Loads and parses the configuration defined in `./app/config.yml`
@@ -15,8 +21,8 @@ function loadConfig() {
   // make sure deployment env variable is set
   if (!process.env.DEPLOYMENT) {
     throw new Error(
-      'You MUST set DEPLOYMENT environment variable with the name' +
-      ' of your deployment before running tests.'
+      'You MUST set DEPLOYMENT environment variable with the name'
+      + ' of your deployment before running tests.'
     );
   }
 
@@ -48,13 +54,48 @@ function loadConfig() {
  * @returns {string} - File path and name of output file (json)
  */
 function templateFile({ inputTemplateFilename, config }) {
-  const inputTemplate = JSON.parse(fs.readFileSync(inputTemplateFilename));
+  const inputTemplate = JSON.parse(fs.readFileSync(inputTemplateFilename, 'utf8'));
   const templatedInput = lodash.merge(lodash.cloneDeep(inputTemplate), config);
   let jsonString = JSON.stringify(templatedInput, null, 2);
   jsonString = jsonString.replace('{{AWS_ACCOUNT_ID}}', config.AWS_ACCOUNT_ID);
   const templatedInputFilename = inputTemplateFilename.replace('.template', '');
   fs.writeFileSync(templatedInputFilename, jsonString);
   return templatedInputFilename;
+}
+
+/**
+ * Upload a file from the test-data package to the S3 test data
+ *
+ * @param {string} file - filename of data to upload
+ * @param {string} bucket - bucket to upload to
+ * @param {string} prefix - S3 folder prefix
+ * @param {boolean} replacePaths - whether to replace test paths in file contents
+ * @returns {Promise<Object>} - promise returned from S3 PUT
+ */
+function uploadTestDataToS3(file, bucket, prefix = 'cumulus-test-data/pdrs', replacePaths = false) {
+  let data = fs.readFileSync(require.resolve(file), 'utf8');
+  if (replacePaths) {
+    data = globalReplace(data, 'cumulus-test-data/pdrs', prefix);
+  }
+  const key = path.basename(file);
+  return s3().putObject({
+    Bucket: bucket,
+    Key: `${prefix}/${key}`,
+    Body: data
+  }).promise();
+}
+
+/**
+ * For the given bucket, upload all the test data files to S3
+ *
+ * @param {string} bucket - S3 bucket
+ * @param {Array<string>} data - list of test data files
+ * @param {string} prefix - S3 folder prefix
+ * @param {boolean} replacePaths - whether to replace test paths in file contents
+ * @returns {Array<Promise>} - responses from S3 upload
+ */
+function uploadTestDataToBucket(bucket, data, prefix, replacePaths) {
+  return Promise.all(data.map((file) => uploadTestDataToS3(file, bucket, prefix, replacePaths)));
 }
 
 /**
@@ -65,15 +106,13 @@ function templateFile({ inputTemplateFilename, config }) {
  * @returns {Promise} undefined
  */
 async function deleteFolder(bucket, folder) {
-  const s3 = new S3();
-
-  const l = await s3.listObjectsV2({
+  const l = await s3().listObjectsV2({
     Bucket: bucket,
     Prefix: folder
   }).promise();
 
   await Promise.all(l.Contents.map((item) =>
-    s3.deleteObject({
+    s3().deleteObject({
       Bucket: bucket,
       Key: item.Key
     }).promise()));
@@ -87,32 +126,68 @@ async function deleteFolder(bucket, folder) {
  */
 function getExecutionUrl(executionArn) {
   const region = process.env.AWS_DEFAULT_REGION || 'us-east-1';
-  return `https://console.aws.amazon.com/states/home?region=${region}` +
-         `#/executions/details/${executionArn}`;
+  return `https://console.aws.amazon.com/states/home?region=${region}`
+         + `#/executions/details/${executionArn}`;
 }
+
 
 /**
- * Redeploy the current Cumulus deployment
+ * Redeploy the current Cumulus deployment.
+ *
+ * Prints '.' per minute while running.  Prints STDOUT from deployCommand
+ * and STDERR if an error occurs.
  *
  * @param {Object} config - configuration object from loadConfig()
- * @returns {undefined} none
+ * @param {int} timeout - Timeout value in minutes
+ * @returns {Promise}
  */
-async function redeploy(config) {
+
+function redeploy(config, timeout) {
   const deployCommand = `./node_modules/.bin/kes  cf deploy --kes-folder app --template node_modules/@cumulus/deployment/app --deployment ${config.deployment} --region us-east-1`;
   console.log(`Redeploying ${config.deployment}`);
-  await exec(deployCommand)
-    .then((result) => {
-      console.log(result.stdout);
-      if (result.error) {
-        console.log(`Deployment error: ${result.error}`);
+
+  let timeoutObject;
+  function timeoutPromise() {
+    return new Promise((resolve, reject) => {
+      const minutes = timeout || 30;
+      let i = 0;
+      function printDots() {
+        console.log('.');
+        if (i < minutes) {
+          i += 1;
+          timeoutObject = setTimeout(printDots, 60000);
+        }
+        else {
+          reject(new Error('Timeout Exceeded'));
+        }
       }
+      printDots();
     });
-  console.log(`Redeploy of ${config.deployment} complete`);
+  }
+
+  async function executionPromise() {
+    let output;
+    try {
+      output = await exec(deployCommand);
+      console.log(output.stdout);
+    }
+    catch (e) {
+      console.log(e.stdout);
+      console.log(e.stderr);
+      throw (e);
+    }
+  }
+
+  return Promise.race([executionPromise(), timeoutPromise()]).then((_) => clearTimeout(timeoutObject));
 }
 
+
 module.exports = {
+  timestampedTestDataPrefix,
   loadConfig,
   templateFile,
+  uploadTestDataToS3,
+  uploadTestDataToBucket,
   deleteFolder,
   getExecutionUrl,
   redeploy